pyproject.toml
[project]
name = "certified-turtles"
version = "0.1.0"
description = "GPTHub / True Tech — backend и интеграция MWS GPT"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ddgs>=9.11.3",
    "fastapi>=0.115.6",
    "python-multipart>=0.0.20",
    "matplotlib>=3.9.0",
    "numpy>=2.0.0",
    "pandas>=2.2.0",
    "openpyxl>=3.1.0",
    "python-pptx>=1.0.2",
    "uvicorn[standard]>=0.32.1",
    "requests>=2.32.0",
]

[project.urls]
Repository = "https://github.com/potamotus/mts-true-tech-2026-certified-turtles"

[build-system]
requires = ["hatchling>=1.26.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/certified_turtles"]

[project.scripts]
mws-gpt = "certified_turtles.mws_gpt.__main__:main"

[project.optional-dependencies]
# pydub на 3.13+: stdlib audioop удалён; на 3.12 пакет не ставим (там audioop ещё есть).
openwebui = [
    "open-webui>=0.6",
    "audioop-lts>=0.2; python_version>='3.13'",
]
# Google Docs API (тулы google_docs_*); JSON ключ service account не в репозитории.
google = [
    "google-api-python-client>=2.160.0",
    "google-auth>=2.35.0",
]
# Изолированный venv: см. externals/requirements-gpt-researcher.txt и scripts/bootstrap_gpt_researcher_venv.sh

[tool.uv]
package = true

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.3",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
